Fixed a bug when first player is making a first move

+++ GoGame.Utility/Constants/StonesConstants.cs
@@ -12,7 +12,7 @@ public static partial class Constants
     public static Brush MouseWhiteStone { get; } = new SolidColorBrush(Color.FromArgb(127, 255, 255, 255));
     public static Brush MouseBlackStone { get; } = new SolidColorBrush(Color.FromArgb(127, 0, 0, 0));
 
-    public static double StoneSize => CellSize;
+    public static int StoneSize => (int)CellSize;
 
     public static Point UndefinedPoint { get; } = new() { X = -1, Y = -1 };
     public static (int I, int J) UndefinedIndexer { get; } = (-1, -1);
